requirements.txt
streamlit
mysql-connector-python
openai==0.28
python-pptx
pillow
python-dotenv